fix(ci): resolve frontend lint warnings
